+++ Packages/KeyboardCore/Sources/KeyboardCore/KeyboardInputSettings.swift
@@ -1,3 +1,4 @@
 public enum KeyboardInputSettingsKey {
     public static let pairedSymbolCompletionEnabled = "paired_symbol_completion_enabled"
+    public static let postCommitContinuationEnabled = "post_commit_continuation_enabled"
 }